added helper for making the complete grades. This is done to allow the process to be executed on the server instead of the client.
